docs/requirements.txt
PyYAML==6.0.1
requests==2.31.0
sphinx==7.2.6
sphinx-carousel==1.2.0
sphinx-copybutton==0.5.2
sphinx-design==0.5.0
sphinx-rtd-theme==2.0.0rc4
